Finished delete

+++ cm-nx-workshop/libs/tourtitan/features/src/lib/features/cars/cars-detail/cars-detail.component.ts
@@ -4,6 +4,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { ICar } from '@cm-nx-workshop/shared/api';
 import { CarsService } from '../cars.service';
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
+import { CarDeleteComponent } from './car-delete/car-delete.component';
 // import { CarDeleteComponent } from './car-delete/car-delete.component';
 
 @Component({
@@ -34,21 +35,35 @@ import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
       });
     }
   
-    deleteCar(): void {
-      if (this.car?.id) {
-        this.carsService.delete(this.car).subscribe({
-          next: () => {
-            console.log('Car deleted successfully');
-            // Optionally, you can navigate to another page or perform additional actions after deletion.
-          },
-          error: (error) => {
-            console.error('Error deleting car:', error);
-          },
-        });
-      } else {
-        console.error('Car id is missing for deletion.');
-      }
+
+  deleteCar(): void {
+    if (this.car) {
+      const modalRef: NgbModalRef = this.modalService.open(CarDeleteComponent, {
+        centered: true,
+        backdrop: false,
+      });
+      modalRef.componentInstance.car = this.car;
+
+      modalRef.componentInstance.confirmDelete.subscribe(() => {
+        if (this.car?.id) {
+          this.carsService.delete(this.car).subscribe({
+            next: () => {
+              console.log('Car deleted successfully');
+              this.router.navigate(['/cars']); 
+            },
+            error: (error) => {
+              console.error('Error deleting car:', error);
+            },
+          });
+        } else {
+          console.error('Car id is missing for deletion.');
+        }
+      });
+    } else {
+      console.error('Car not found.');
     }
+  }
+
   
     ngOnDestroy(): void {
       if (this.subscription) this.subscription.unsubscribe();
